requirements.txt
fastapi
uvicorn
supabase
python-dotenv
langchain-google-genai
pydantic